--- src/main/java/com/example/qsproject/qsproject/CourseRepository.java
@@ -1,8 +1,0 @@
-package com.example.qsproject.qsproject;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-// Isto faz com que o User tenha os métodos CRUD (create ,read ,update,delete)
-public interface CourseRepository extends JpaRepository<Course,Integer> {
-}
