+++ src/main/java/com/example/qsproject/qsproject/Implementation/SubjectImpl.java
@@ -1,4 +1,5 @@
 package com.example.qsproject.qsproject.Implementation;
 
 public class SubjectImpl {
+
 }
