+++ src/main/java/com/example/qsproject/qsproject/controllers/SubjectController.java
@@ -1,4 +1,5 @@
 package com.example.qsproject.qsproject.controllers;
 
 public class SubjectController {
+
 }
