+++ src/main/java/com/example/qsproject/qsproject/dtos/CourseDto.java
@@ -6,16 +6,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
+import java.util.ArrayList;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-
 public class CourseDto {
-    private long courseId;
-    private String courseName;
-    private List<Subject> subjects;
+    private long CourseId;
+    private String CourseName;
+    private ArrayList<Subject> subjects;
 }
-
