+++ src/main/java/com/example/qsproject/qsproject/mappers/ClassroomMapper.java
@@ -2,12 +2,6 @@ package com.example.qsproject.qsproject.mappers;
 
 import com.example.qsproject.qsproject.*;
 import com.example.qsproject.qsproject.dtos.ClassroomDto;
-import com.example.qsproject.qsproject.dtos.CourseDto;
-import com.example.qsproject.qsproject.dtos.SubjectDto;
-import com.example.qsproject.qsproject.dtos.UsersDto;
-
-import java.sql.Date;
-import java.util.List;
 
 public class ClassroomMapper {
