+++ src/main/java/com/example/qsproject/qsproject/mappers/CourseMapper.java
@@ -1,11 +1,7 @@
 package com.example.qsproject.qsproject.mappers;
 
 import com.example.qsproject.qsproject.Course;
-import com.example.qsproject.qsproject.Subject;
-import com.example.qsproject.qsproject.User;
 import com.example.qsproject.qsproject.dtos.CourseDto;
-import com.example.qsproject.qsproject.dtos.SubjectDto;
-import com.example.qsproject.qsproject.dtos.UsersDto;
 
 public class CourseMapper {
 
@@ -17,15 +13,11 @@ public class CourseMapper {
         );
     }
 
-    public static Course mapToCourse(CourseDto courseDto){
+    public static Course mapToCourse(CourseDto courseDto) {
         return new Course(
                 courseDto.getCourseId(),
                 courseDto.getCourseName(),
                 courseDto.getSubjects()
         );
     }
-
-
 }
-
-
