+++ src/main/java/com/example/qsproject/qsproject/repositories/CourseRepository.java
@@ -1,4 +1,9 @@
 package com.example.qsproject.qsproject.repositories;
 
-public interface CourseRepository {
+import com.example.qsproject.qsproject.Course;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+// Isto faz com que o User tenha os métodos CRUD (create ,read ,update,delete)
+public interface CourseRepository extends JpaRepository<Course,Integer> {
 }
