+++ src/main/java/com/example/qsproject/qsproject/services/SubjectServices.java
@@ -1,8 +1,6 @@
 package com.example.qsproject.qsproject.services;
 
-import com.example.qsproject.qsproject.dtos.CourseDto;
 import com.example.qsproject.qsproject.dtos.SubjectDto;
-import com.example.qsproject.qsproject.dtos.UsersDto;
 
 import java.util.List;
